Restructure anti-patterns to seperate package
