+++ sqleyes/detector/antipatterns/abstract_base_class.py
