+++ sqleyes/detector/antipatterns/fear_of_the_unknown.py
@@ -1,6 +1,6 @@
 """Fear of the Unknown anti-pattern detector class"""
 import re
-from sqleyes.detector.abstract_ap import APDetector
+from sqleyes.detector.antipatterns.abstract_base_class import APDetector
 from sqleyes.detector.detector_output import DetectorOutput
