+++ sqleyes/detector/antipatterns/implicit_columns.py
@@ -1,6 +1,6 @@
 """Implicit Columns anti-pattern detector class"""
 import re
-from sqleyes.detector.abstract_ap import APDetector
+from sqleyes.detector.antipatterns.abstract_base_class import APDetector
 from sqleyes.detector.detector_output import DetectorOutput
